testing

+++ src/ExternalCSources/fec/libmoepgf/gf256_neon.h
@@ -24,10 +24,10 @@ xorr_neon_64(uint8_t *region1, const uint8_t *region2, size_t length)
     register uint8x8_t in, out;
 
     for (end=region1+length; region1<end; region1+=8, region2+=8) {
-        in  = vld1_u8((const uint64_t *)region2);
-        out = vld1_u8((const uint64_t *)region1);
+        in  = vld1_u8((const uint8_t *)region2);
+        out = vld1_u8((const uint8_t *)region1);
         out = veor_u8(in, out);
-        vst1_u8((uint64_t *)region1, out);
+        vst1_u8((uint8_t *)region1, out);
     }
     std::cout<<"Y neon\n";
 }
